+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.3.0 - 12/04/2011
+
+- Renamed `Deputy::get_title` to `Deputy::title` and made a getter/setter
+- Renamed `Deputy::get_uri` to `Deputy::uri` and made a getter/setter
+- Refactored `Deputy::is_visible` to a getter/setter
+- Added `Deputy_Resource::segment` for individual Resource URI segment
+- Added `Deputy_Resource::set_meta` and `Deputy_Resource::get_meta` for custom meta data.
+- Updated User Guide documentation
+- Added `Kohana_Deputy_Resource_Test::test_meta` and updated tests to use new naming.
+- All tests pass: OK (11 tests, 73 assertions)
+
 # 0.2.0 - 10/23/2011
 
 - Upgraded to support Kohana 3.2
@@ -10,4 +21,4 @@
 - Initial release of Deputy
 - Role and Resource support
 - User Guide documentation
-- Unit Test coverage with
+- Unit Test coverage